+++ asus-nb/src/profile.rs
@@ -76,6 +76,8 @@ fn parse_fan_curve(data: &str) -> Result<Curve, String> {
 pub struct ProfileCommand {
     #[options(help = "print help message")]
     help: bool,
+    #[options(help = "toggle to next profile in list")]
+    pub next: bool,
     #[options(help = "create the profile if it doesn't exist")]
     pub create: bool,